+++ core/src/main/kotlin/io/github/shaksternano/borgar/core/collect/SizedIterable.kt
@@ -1,5 +1,5 @@
 package io.github.shaksternano.borgar.core.collect
 
 interface SizedIterable<T> : Iterable<T> {
-    val size: Long
+    val size: Int
 }
